# Add request duration to diagnostic metadata