As titled.
Now 100% convinced this should live in the transport, as it can be obtained with observability events and it causes a minor slowdown in performances.

Code used for benchmarks:
```js
'use strict'

const http = require('http')

const server = http.createServer((req, res) => {
  res.writeHead(200, {
    'content-type': 'application/json'
  })
  res.end('{"hello":"world"}')
})

server.listen(3000, async () => {
  console.log('listening')
})
```

```js
'use strict'

const { Transport, WeightedConnectionPool, UndiciConnection } = require('./')

const pool = new WeightedConnectionPool({ Connection: UndiciConnection })
pool.addConnection('http://localhost:3000')
const transport = new Transport({ connectionPool: pool })

async function run () {
  for (let i = 0; i < 1000; i++) {
    await transport.request({ method: 'GET', path: '/' })
  }
}

run().catch(err => {
  console.log(err)
  process.exit(1)
})
```

Results:
```
# this branch
  Time (mean ± σ):     438.2 ms ±  51.3 ms    [User: 470.0 ms, System: 83.4 ms]
  Range (min … max):   389.0 ms … 521.6 ms    10 runs

# main branch
Benchmark 1: node bench.js
  Time (mean ± σ):     405.5 ms ±  35.3 ms    [User: 451.2 ms, System: 77.6 ms]
  Range (min … max):   366.6 ms … 478.9 ms    10 runs
```
The benchmark has been taken with [hyperfine](https://github.com/sharkdp/hyperfine).