@mshustov 2022-02-22
Another option is to introduce a flag on the `transport` level to measure request duration. However, it comes with the cost of increasing API surface. WDYT @delvedor?